moving helper functions to separate file
