+++ client/src/Root.js
@@ -4,7 +4,7 @@ import { BrowserRouter as Router, Route } from 'react-router-dom';
 
 import Main from './containers/Main';
 import Garage from './containers/Garage';
-import Game from './containers/Play';
+import Game from './containers/Play/Play';
 import Shop from './containers/Shop';
 
 function Root() {
