+++ client/src/containers/Play/GameUI.scss
@@ -1,5 +1,5 @@
-@import './../styles/Variables';
-@import './../styles/Mixins';
+@import './../../styles/Variables';
+@import './../../styles/Mixins';
 
 .DisplayQuote-previewQuote {
   border-radius: 1rem;
