+++ client/src/containers/Play/Play.js
@@ -1,70 +1,13 @@
 import React, { Component, Fragment } from 'react';
-import { withRouter } from 'react-router-dom';
+import socketIOClient from 'socket.io-client';
 import './GameUI.scss';
 
-import tokenValidationHelper from './../lib/tokenValidationHelper';
-// import DisplayQuote from './GameUI/GameUI';
-import Background from './../components/Background/Background';
-import DisplayQuoteArea from './../components/Quote/Quote';
-import CarWPMGauge from './../components/Guages/WPMGuage';
-import Minimap from './../components/Minimap/Minimap';
-import DisplayQuoteInput from './../components/GameInput/GameInput';
-import NosGauge from './../components/Guages/NOSGuage';
-import socketIOClient from 'socket.io-client';
-import StartGameButton from './../components/StartGameButton/StartGameButton';
-import LeaderboardModal from './../components/LeaderboardModal/LeaderboardModal';
-import Header from './../components/Header/Header';
+import tokenValidationHelper from './../../lib/tokenValidationHelper';
 
-const renderGame = props => {
-  const countdown = props.countdown ? (
-    <h1>{props.countdownCount}</h1>
-  ) : (
-    <StartGameButton emitStart={props.onEmitStart} />
-  );
+import { renderGame } from './../../lib/playHelpers';
 
-  const gameStart = props.timerStart ? (
-    <Fragment>
-      <DisplayQuoteArea
-        completed={props.wordsCompleted}
-        input={props.userInput}
-        currentWord={props.words[props.index]}
-        remaining={props.remainingPhrase}
-      />
-      <DisplayQuoteInput onUserInputChange={props.onUserInputChange} />
-    </Fragment>
-  ) : (
-    <DisplayQuoteArea
-      completed={''}
-      input={''}
-      currentWord={''}
-      remaining={'Please wait for the game to start'}
-    />
-  );
-
-  return props.timerFinished ? (
-    <Fragment>
-      <LeaderboardModal
-        leaderboard={props.leaderboard}
-        placings={props.placings}
-      />
-    </Fragment>
-  ) : (
-    <div className="DisplayQuoteUI-container">
-      <CarWPMGauge
-        wpm={props.wpm}
-        second={props.sec}
-        char={props.char}
-        socket={props.socket}
-      />
-      <div className="DisplayQuote-container">
-        {countdown}
-        <Minimap playerProgress={props.playerProgress} />
-        {gameStart}
-      </div>
-      <NosGauge position={props.position} />
-    </div>
-  );
-};
+import Background from './../../components/Background/Background';
+import Header from './../../components/Header/Header';
 
 class PlayGameLogic extends Component {
   constructor() {
@@ -73,7 +16,7 @@ class PlayGameLogic extends Component {
     this.state = {
       user: {
         username: '',
-        car: '',
+        car: ''
       },
       countdownCount: 5,
       countdown: false,
@@ -108,7 +51,6 @@ class PlayGameLogic extends Component {
   }
 
   async componentDidMount() {
-
     // Set user in state via localStorage
     const user = await tokenValidationHelper();
 
@@ -120,10 +62,7 @@ class PlayGameLogic extends Component {
     const socket = socketIOClient(endpoint);
 
     socket.on('connect', () => {
-      socket.emit(
-        'user-update',
-        JSON.stringify({ user: this.state.user })
-      );
+      socket.emit('user-update', JSON.stringify({ user: this.state.user }));
 
       this.setState({
         socket
@@ -188,12 +127,11 @@ class PlayGameLogic extends Component {
 
         // Set completed boolean for the leaderboard.
         if (leaderboard[message.socketId].completion === true) {
-          leaderboard[message.socketId].completed = true
+          leaderboard[message.socketId].completed = true;
         } else {
-          leaderboard[message.socketId].completed = false
+          leaderboard[message.socketId].completed = false;
         }
       }
-
 
       this.setState({ leaderboard });
 
@@ -266,7 +204,12 @@ class PlayGameLogic extends Component {
 
     socket.on('disconnect', () => {
       alert('Please reload your page');
-      this.setState({ leaderboard: {}, placings: [], progress: 0, carPositioning: {} });
+      this.setState({
+        leaderboard: {},
+        placings: [],
+        progress: 0,
+        carPositioning: {}
+      });
     });
   }
